fix: wrong type in migration

+++ migration/src/m20220826_174342_create_table.rs
@@ -34,7 +34,7 @@ fn course() -> TableCreateStatement {
                 .custom(Alias::new("LONGTEXT"))
                 .not_null(),
         )
-        .col(ColumnDef::new(Alias::new("credit")).float().not_null())
+        .col(ColumnDef::new(Alias::new("credit")).double().not_null())
         .col(
             ColumnDef::new(Alias::new("department"))
                 .custom(Alias::new("LONGTEXT"))
